Create crossroad-html.html
This will be my home/landing page. From here the user will have access to all the different portions of the original website but in a much clearer and intuitive manner. Each bubble will expand to show the area of the website it will take you to and will also expand on hover.