Create education-html.html
This is the new education section for the website and it will have its own html file now. I expanded the page a bit to allow for me to fill out more of my educational background and to have a better flow to this section of the website. Now there is room to expand and add mroe information instead of being limited to the 100vh section in main-html.