Update and rename main-css.css to work-scss.scss
Changed the main-scss file to become the work-scss file to display my previous and current work experience and style it as its own page. So far only have a landing page image and a title, will be adding in to different divs that will open on :hover and display information on my work experience and have links to better illustrate what I have done and am currently doing.